changed movement to float

+++ src/main/java/arkiGame/components/SampleComponent.java
@@ -3,7 +3,7 @@ package arkiGame.components;
 import l0raxeo.arki.engine.components.Component;
 import l0raxeo.arki.engine.components.collisionComponents.RigidBody;
 import l0raxeo.arki.engine.input.mouse.MouseManager;
-import org.joml.Vector2i;
+import org.joml.Vector2f;
 
 public class SampleComponent extends Component
 {
@@ -11,7 +11,7 @@ public class SampleComponent extends Component
     @Override
     public void update(double dt) {
         if (MouseManager.hasPressedInput())
-            gameObject.getComponent(RigidBody.class).addForce(new Vector2i(0, -10));
+            gameObject.getComponent(RigidBody.class).addForce(new Vector2f(0, -10));
 
         gameObject.transform.rotation++;
     }
