+++ src/main/java/l0raxeo/arki/engine/components/collisionComponents/Bounds.java
@@ -4,6 +4,7 @@ import l0raxeo.arki.engine.collision.Collision;
 import l0raxeo.arki.engine.components.Component;
 import l0raxeo.arki.engine.objects.GameObject;
 import l0raxeo.arki.renderer.AppWindow;
+import org.joml.Vector2f;
 import org.joml.Vector2i;
 
 import java.awt.*;
@@ -27,7 +28,7 @@ public abstract class Bounds extends Component
         bounds.setRect(gameObject.transform.getScreenPosition().x, gameObject.transform.getScreenPosition().y, gameObject.transform.scale.x, gameObject.transform.scale.y);
     }
 
-    public List<Collision> findGameObjectsInPath(Vector2i velocity)
+    public List<Collision> findGameObjectsInPath(Vector2f velocity)
     {
         List<Collision> result = new ArrayList<>();
