+++ src/main/java/l0raxeo/arki/engine/objects/prefabs/Prefabs.java
@@ -3,8 +3,8 @@ package l0raxeo.arki.engine.objects.prefabs;
 import l0raxeo.arki.engine.components.Component;
 import l0raxeo.arki.engine.objects.GameObject;
 import l0raxeo.arki.engine.objects.Transform;
-import org.joml.Vector2i;
-import org.joml.Vector3i;
+import org.joml.Vector2f;
+import org.joml.Vector3f;
 
 public class Prefabs {
 
@@ -13,7 +13,7 @@ public class Prefabs {
 
     }
 
-    public static GameObject generate(String name, Vector3i pos, Vector2i size, Component... comps)
+    public static GameObject generate(String name, Vector3f pos, Vector2f size, Component... comps)
     {
         GameObject go = new GameObject(name, new Transform(pos.add(0, size.y, 0), size));
 
@@ -24,7 +24,7 @@ public class Prefabs {
         return go;
     }
 
-    public static GameObject generate(String name, Vector3i pos, Vector2i size, float rotation, Component... comps)
+    public static GameObject generate(String name, Vector3f pos, Vector2f size, float rotation, Component... comps)
     {
         GameObject go = new GameObject(name, new Transform(pos.add(0, size.y, 0), size, rotation));
